backend/requirements.txt
asgiref==3.4.1
Django==3.2.6
djangorestframework==3.12.4
pytz==2021.1
sqlparse==0.4.2
auth0-python==3.16.2
django-environ==0.5.0
psycopg2-binary==2.9.1
django-cors-headers
django-nose
django_createsuperuser
pdfkit
